+++ src/lib.rs
@@ -55,7 +55,6 @@ unstable_module! {
     pub mod messagebuffer;
     pub mod mutex;
     pub mod prioritydataqueue;
-    #[cfg(any(feature = "fmp3", feature = "solid_fmp3", feature = "none"))]
     pub mod processor;
     pub mod semaphore;
     pub mod task;