+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 - **Changed:** `itron::task::Builder::{stack → stack_size}`
+- **Changed:** `itron::{processor, task::{Task::activate_on, Builder::initial_processor}}`, etc. are now exposed on uniprocessor kernels
 - **Added:** `itron::abi::{acre_cyc, acre_alm, acre_isr}`
 - **Added:** `itron::abi::acre_mpf` when `cfg(all(feature = "solid_asp3", feature = "dcre"))`
 - **Added:** `itron::abi::exd_tsk`